+++ simply-synonyms/src/api/synonyms.js
@@ -43,14 +43,14 @@ const synonymsApi = {
     const synonymRequestPromise = GET(`get-dictionary-data?word=${word}`)
       .then(response => response.json())
       .then(data => {
-        // if (!!this.idToken && !userCancelledRequest) GET('update-user-stats?word-data=1') // Increment user's synonym counters once they recieve the synonym data
-        //   .then(({ status }) => {
-        //     if (status === 401) {
-        //       browser.runtime.sendMessage(null, { action: 'refreshIdToken' }, {}, t => {
-        //         this.idToken = t
-        //       }) // Check for a token refresh and update token when fetch is unauthorized
-        //     }
-        //   })
+        if (!!this.idToken && !userCancelledRequest) GET('update-user-stats?worddata=1') // Increment user's synonym counters once they recieve the synonym data
+          .then(({ status }) => {
+            if (status === 401) {
+              browser.runtime.sendMessage(null, { action: 'refreshIdToken' }, {}, t => {
+                this.idToken = t
+              }) // Check for a token refresh and update token when fetch is unauthorized
+            }
+          })
         return data
       })